+++ hokusai/commands/logs.py
@@ -4,8 +4,8 @@ from hokusai.lib.common import shout, shout_concurrent
 from hokusai.services.kubectl import Kubectl
 
 @command
-def logs(context, timestamps, follow, tail):
-  kctl = Kubectl(context)
+def logs(context, timestamps, follow, tail, namespace=None):
+  kctl = Kubectl(context, namespace=namespace)
 
   opts = ''
   if timestamps:
@@ -15,7 +15,7 @@ def logs(context, timestamps, follow, tail):
   if tail:
     opts += " --tail=%s" % tail
 
-  pods = kctl.get_object('pod', selector="app=%s,layer=application" % config.project_name)
+  pods = kctl.get_objects('pod', selector="app=%s,layer=application" % config.project_name)
   pods = filter(lambda pod: pod['status']['phase'] == 'Running', pods)
   containers = []