+++ hokusai/services/deployment.py
@@ -9,19 +9,31 @@ from hokusai.services.command_runner import CommandRunner
 from hokusai.lib.exceptions import HokusaiError
 
 class Deployment(object):
-  def __init__(self, context):
+  def __init__(self, context, deployment_name=None, namespace=None):
     self.context = context
-    self.kctl = Kubectl(self.context)
+    self.namespace = namespace
+    self.kctl = Kubectl(self.context, namespace=namespace)
     self.ecr = ECR()
-    self.cache = self.kctl.get_object('deployment', selector="app=%s,layer=application" % config.project_name)
+    if deployment_name:
+      self.cache = [self.kctl.get_object("deployment %s" % deployment_name)]
+    else:
+      self.cache = self.kctl.get_objects('deployment', selector="app=%s,layer=application" % config.project_name)
 
-  def update(self, tag, constraint, git_remote):
+  def update(self, tag, constraint, git_remote, resolve_tag_sha1=True):
     if not self.ecr.project_repo_exists():
       raise HokusaiError("Project repo does not exist.  Aborting.")
 
-    print_green("Deploying %s to %s..." % (tag, self.context))
+    if resolve_tag_sha1:
+      tag = self.ecr.find_git_sha1_image_tag(tag)
+      if tag is None:
+        raise HokusaiError("Could not find a git SHA1 for tag %s.  Aborting." % tag)
 
-    if self.context != tag:
+    if self.namespace is None:
+      print_green("Deploying %s to %s..." % (tag, self.context))
+    else:
+      print_green("Deploying %s to %s/%s..." % (tag, self.context, self.namespace))
+
+    if self.namespace is None:
       self.ecr.retag(tag, self.context)
       print_green("Updated tag %s -> %s" % (tag, self.context))
 
@@ -36,8 +48,8 @@ class Deployment(object):
         shout("git push --force %s --tags" % git_remote, print_output=True)
 
     if config.pre_deploy is not None:
-      print_green("Running pre-deploy hook '%s' on %s..." % (config.pre_deploy, self.context))
-      return_code = CommandRunner(self.context).run(tag, config.pre_deploy, constraint=constraint)
+      print_green("Running pre-deploy hook '%s'..." % config.pre_deploy)
+      return_code = CommandRunner(self.context, namespace=self.namespace).run(tag, config.pre_deploy, constraint=constraint)
       if return_code:
         raise HokusaiError("Pre-deploy hook failed with return code %s" % return_code, return_code=return_code)
 
@@ -69,8 +81,8 @@ class Deployment(object):
       raise HokusaiError("Deployment failed!", return_code=return_code)
 
     if config.post_deploy is not None:
-      print_green("Running post-deploy hook '%s' on %s..." % (config.post_deploy, self.context))
-      return_code = CommandRunner(self.context).run(tag, config.post_deploy, constraint=constraint)
+      print_green("Running post-deploy hook '%s'..." % config.post_deploy)
+      return_code = CommandRunner(self.context, namespace=self.namespace).run(tag, config.post_deploy, constraint=constraint)
       if return_code:
         raise HokusaiError("Post-deploy hook failed with return code %s" % return_code, return_code=return_code)